+++ src/editor/fileIcons/index.js
@@ -1,0 +1,5 @@
+// Because the sidebar also use the file icons, So I put this file out of floatBox directory.
+import 'file-icons-js/css/style.css'
+import fileIcons from 'file-icons-js'
+
+export default fileIcons